+++ pyinterpolate/distance/distance.py
@@ -218,18 +218,20 @@ def _calculate_block_to_block_distance(block_1: np.ndarray, block_2: np.ndarray)
     return distances_sum
 
 
-def _calc_angle_between_points(v1, v2):
-    ang1 = np.arctan2(*v1[::-1])
-    ang2 = np.arctan2(v2[:, 1], v2[:, 0])
+def _calc_angle_between_points(v1, v2, origin):
+    ang1 = np.arctan2(v1[1] - origin[1], v1[0] - origin[0])
+    ang2 = np.arctan2(v2[:, 1] - origin[1], v2[:, 0] - origin[0])
     return np.rad2deg((ang1 - ang2) % (2 * np.pi))
 
 
-def _calc_angle_from_origin(vec):
-    ang = np.arctan2(vec[:, 1], vec[:, 0])
+def _calc_angle_from_origin(vec, origin):
+    ys = vec[:, 1] - origin[1]
+    xs = vec[:, 0] - origin[0]
+    ang = np.arctan2(ys, xs)
     return np.rad2deg(ang % (2 * np.pi))
 
 
-def calc_angles(points_b: Iterable, point_a: Iterable = None):
+def calc_angles(points_b: Iterable, point_a: Iterable = None, origin: Iterable = None):
     """
     Function calculates angles between points and origin or between vectors from origin to points and a vector from
     a specific point to origin.
@@ -240,24 +242,34 @@ def calc_angles(points_b: Iterable, point_a: Iterable = None):
         Other point coordinates.
 
     point_a : Iterable
-        The point coordinates, default is equal to (0, 0) - origin.
+        The point coordinates, default is equal to (0, 0).
+
+    origin : Iterable
+        The origin coordinates, default is (0, 0).
 
     Returns
     -------
     angles : numpy array
-        Angles from the ``points_b`` to origin, or angles beteen vectors ``points_b`` to origin and ``point_a``
+        Angles from the ``points_b`` to origin, or angles between vectors ``points_b`` to origin and ``point_a``
         to origin.
     """
 
+    if origin is None:
+        origin = np.array((0, 0))
+    else:
+        if not isinstance(origin, np.ndarray):
+            origin = np.array(origin)
+
     if not isinstance(points_b, np.ndarray):
         points_b = np.array(points_b)
+
     if len(points_b.shape) == 1:
         points_b = points_b[np.newaxis, ...]
 
     if point_a is None:
-        angles = _calc_angle_from_origin(points_b)
+        angles = _calc_angle_from_origin(points_b, origin)
     else:
-        angles = _calc_angle_between_points(point_a, points_b)
+        angles = _calc_angle_between_points(point_a, points_b, origin)
 
     return angles
 
@@ -290,12 +302,12 @@ def calc_point_to_point_distance(points_a, points_b=None):
 
 def calculate_angular_distance(angles: np.ndarray, expected_direction: float) -> np.ndarray:
     """
-    Function calculates minimal angle between one vector and other vectors.
+    Function calculates minimal direction between one vector and other vectors.
 
     Parameters
     ----------
     angles : numpy array
-        The array with the angle to the origin of each point.
+        The array with the direction to the origin of each point.
 
     expected_direction : float
         The variogram direction in degrees.
@@ -303,7 +315,7 @@ def calculate_angular_distance(angles: np.ndarray, expected_direction: float) ->
     Returns
     -------
     angular_distances : numpy array
-        Minimal angle from ``expected_direction`` to other angles.
+        Minimal direction from ``expected_direction`` to other angles.
     """
 
     # We should select angles equal to the expected direction