+++ pyinterpolate/kriging/models/block/area_to_point_poisson_kriging.py
@@ -21,7 +21,7 @@ from pyinterpolate.kriging.models.block.weight import WeightedBlock2BlockSemivar
 from pyinterpolate.processing.preprocessing.blocks import Blocks, PointSupport
 from pyinterpolate.processing.select_values import select_poisson_kriging_data, prepare_pk_known_areas, \
     get_aggregated_point_support_values
-from pyinterpolate.processing.transform.transform import transform_ps_to_dict, get_areal_values_from_agg
+from pyinterpolate.processing.transform.transform import transform_ps_to_dict, get_areal_values_from_agg, sem_to_cov
 from pyinterpolate.variogram import TheoreticalVariogram
 
 
@@ -107,6 +107,9 @@ def area_to_point_pk(semivariogram_model: TheoreticalVariogram,
     else:
         rng = max_range
 
+    # Get sill to calc cov
+    sill = semivariogram_model.sill
+
     kriging_data = select_poisson_kriging_data(
         u_block_centroid=unknown_block,
         u_point_support=unknown_block_point_support,
@@ -130,7 +133,13 @@ def area_to_point_pk(semivariogram_model: TheoreticalVariogram,
     #      unknown point 2 (n+1) semivariance against point support from block 2,
     #      ...],
     # )
-    avg_b2p_semivariances = add_ones(b2p_semivariance.calculate_average_semivariance(kriging_data))
+    # Transform to covariances
+    avg_b2p_covariances = add_ones(
+        sem_to_cov(
+            b2p_semivariance.calculate_average_semivariance(kriging_data),
+            sill
+        )
+    )
 
     # {(known block id a, known block id b): [pt a val, pt b val, distance between points]}
     distances_between_known_areas = prepare_pk_known_areas(dps, prepared_ids)
@@ -147,6 +156,9 @@ def area_to_point_pk(semivariogram_model: TheoreticalVariogram,
         predicted.append(row)
 
     predicted = np.array(predicted)
+
+    # Transform to covariances
+    predicted = sem_to_cov(predicted, sill)
 
     # Add diagonal weights
     values = get_areal_values_from_agg(blocks, prepared_ids)
@@ -165,7 +177,7 @@ def area_to_point_pk(semivariogram_model: TheoreticalVariogram,
     # Each point of unknown area represents different kriging system
 
     # Transpose points matrix
-    transposed = avg_b2p_semivariances.T
+    transposed = avg_b2p_covariances.T
 
     predicted_points = []