+++ changelog.rst
@@ -6,6 +6,25 @@ Pyinterpolate is the Python library for **geostatistics** and **spatial statisti
 Changes by date
 ===============
 
+2023-01-16
+----------
+
+**version 0.3.6**
+
+* (enhancement) Directional Centroid-based Poisson Kriging,
+* (debug) Added origin (unknown point) to calculate directional Kriging and directional Centroid-based Poisson Kriging,
+* (docs) Directional Ordinary Kriging tutorial,
+* (engancement) logging of area to area PK function,
+* (enhancement) `tests` package moved outside the main package,
+* (feature) ordinary kriging from covariance terms,
+* (feature) area-to-area PK from covariance terms,
+* (debug) area-to-area PK debugged,
+* (feature) area-to-point PK from covariance terms,
+* (debug) area-to-point PK debugged,
+* (feature) centroid-based PK from covariance terms,
+* (debug) centroid-based PK debugged.
+
+
 2022-11-05
 ----------